Refactor of news and stock models to support separate storage
Because we are going to be retrieving data using user credentials, we want
to store that data separately in the database rather than sharing the stocks
across users. Change the models around to allow this, but still allow news
items to be shared across users as that information is free to retrieve and
it doesn't make sense to duplicate it. Also set the stage for smarter
retrieval of news information by storing timestamps for each news item in a
friendly format.