Data storage and refresh refactor
This is the start to breaking out all of our refreshes of data to a more
granular level. Instead of trying to use one date to track them all, use a
bunch of different last update fields on our main stock objects to keep
things straight. Rename a couple fields as well. This does not fully bring
us to the point where we can remove the user_id from the stocks just yet,
however.